% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_GEO_data.r
\name{get_GEO_data}
\alias{get_GEO_data}
\title{Get GEO data from ECDC GIS server}
\usage{
get_GEO_data(layer, STAT_LEVL = c(0), FIELDS = "GEO_ID", isValid = 1)
}
\arguments{
\item{layer}{Select the layer for Hybrid Layer service: 0 for Points, 1 for Polygons.}

\item{STAT_LEVL}{Which values: 0,10=Country level(0 for EU/EEA and candidates, 10 for the rest of world); 
1,2,3 = NUTS levels; 11,12 = GAUL levels; 21,22,23,24 = GADM levels. Defaults to 0 for EU/EEA.}

\item{FIELDS}{= Which "additional" field to retrieve from the hybrid layer, defaults to "GEO_ID". 
"GEO_NAME" also available.}

\item{isValid}{isValid=1 removes the obsolete NUTS regions.}
}
\description{
This function gets the geodata from ECDC GIS server into R SpatialpolygonsDataFrame or SpatialpointsDataFrame. 
Gives a warning when closing the connection.
}
\examples{
# get EU/EEA country polygons
plgs <- get_GEO_data(layer = 1)

# get EU/EEA country points
points <- get_GEO_data(layer = 0)

# use the geodata
plot(points, col = "red")
plot(plgs, add = TRUE)
}
\author{
Tommi Karki; with thanks to Silviu Ionescu.
}
