% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrepMap.R
\name{PrepMap}
\alias{PrepMap}
\title{Prepare your data and spatial data for SurvMapper (ggplot)}
\usage{
PrepMap(data, geo)
}
\arguments{
\item{data}{Your data that you want to include in the map, including the variable GEO_ID with the correct codes}

\item{geo}{Your geometry data, either SpatialPolygonsDataFrame or SpatialPointsDataFrame with variable GEO_ID}
}
\description{
Prepare your data and spatial data for SurvMapper (ggplot)
}
\examples{

# load the included dummy data
load(system.file("extdata", "dummy_data.rds", package = "SurvMaps"))
# load the included EU/EEA SpatialPolygonsDataframe (includes Asia and Africa for background)
load(system.file("extdata", "EU_AFR_AS_plgs.rds", package = "SurvMaps"))

mymap <- PrepMap(data = dummy_data , geo = plg_map)
}
\keyword{map}
