% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_GEO_data.r
\name{get_GEO_data}
\alias{get_GEO_data}
\title{Get GEO data from ECDC GIS server}
\usage{
get_GEO_data(layer, STAT_LEVL = c(0), FIELDS = "GEO_ID", isValid = 1)
}
\arguments{
\item{layer}{Select the layer for Hybrid Layer service: 0 for Points, 1 for Polygons.}

\item{STAT_LEVL}{Which data level to choose; see details. Defaults to 0 for EU/EEA. Give a single value or vector.}

\item{FIELDS}{Which data fields to retrieve from the hybrid layer; see details. Defaults to "GEO_ID".
Give a single value or character vector of relevant field names.}

\item{isValid}{isValid=1 removes the obsolete NUTS regions.}
}
\description{
This function gets the geodata from ECDC GIS server into R SpatialpolygonsDataFrame or SpatialpointsDataFrame.
}
\details{
STAT_LEVL; 0 for EU/EEA and candidate country level, 10 for the rest of world country level. 
1,2,3 = NUTS levels; 11,12 = GAUL levels; 21,22,23,24 = GADM levels. 
FIELDS; for additional fields to retrieve from the hybrid lay, see:
https://gis.ecdc.europa.eu/public/rest/services/UtilitiesServices/EMMA_GoMap_GC_HL_V1_3_WebMerc/MapServer/
}
\examples{
# get EU/EEA country polygons
plgs <- get_GEO_data(layer = 1)

# get EU/EEA country points
points <- get_GEO_data(layer = 0)

# use the geodata
plot(points, col = "red")
plot(plgs, add = TRUE)

# get all the world country polygons and some additional fields
plgs <- get_GEO_data(layer = 1, FIELDS = c("GEO_ID", "GEO_NAME", "CONTINENT"), STAT_LEVL=c(0,10))
}
\author{
Tommi Karki, with thanks to Silviu Ionescu.
}
